Split up rewind_v1_5 gradle project into more subprojects
