+++ rewind_v1_5/rewind_v1_5-common/build.gradle
@@ -1,0 +1,9 @@
+
+plugins {
+	id "java-library"
+}
+
+dependencies {
+	compileOnly project(":api")
+	compileOnly(libs.slf4j)
+}
