+++ rewind_v1_5/rewind_v1_5-common/src/main/java/net/lax1dude/eaglercraft/backend/rewind_v1_5/adapter/IRewindLogger.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2025 lax1dude. All Rights Reserved.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+ * IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+ * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ * 
+ */
+
+package net.lax1dude.eaglercraft.backend.rewind_v1_5.adapter;
+
+public interface IRewindLogger {
+
+	void info(String msg);
+
+	void info(String msg, Throwable thrown);
+
+	void warn(String msg);
+
+	void warn(String msg, Throwable thrown);
+
+	void error(String msg);
+
+	void error(String msg, Throwable thrown);
+
+	IRewindSubLogger createSubLogger(String name);
+
+	public interface IRewindSubLogger extends IRewindLogger {
+
+		IRewindLogger getParent();
+
+		String getName();
+
+		void setName(String name);
+
+	}
+
+}
