+++ rewind_v1_5/rewind_v1_5-common/src/main/java/net/lax1dude/eaglercraft/backend/rewind_v1_5/adapter/IRewindPlatform.java
@@ -14,32 +14,10 @@
  * 
  */
 
-package net.lax1dude.eaglercraft.backend.rewind_v1_5.base;
+package net.lax1dude.eaglercraft.backend.rewind_v1_5.adapter;
 
-public interface IRewindLogger {
+public interface IRewindPlatform<PlayerObject> {
 
-	void info(String msg);
-
-	void info(String msg, Throwable thrown);
-
-	void warn(String msg);
-
-	void warn(String msg, Throwable thrown);
-
-	void error(String msg);
-
-	void error(String msg, Throwable thrown);
-
-	IRewindSubLogger createSubLogger(String name);
-
-	public interface IRewindSubLogger extends IRewindLogger {
-
-		IRewindLogger getParent();
-
-		String getName();
-
-		void setName(String name);
-
-	}
+	IRewindLogger logger();
 
 }
