+++ rewind_v1_5/rewind_v1_5-platform-bukkit/build.gradle
@@ -1,0 +1,12 @@
+
+plugins {
+	id "java-library"
+}
+
+dependencies {
+	implementation project(":rewind_v1_5:rewind_v1_5-common")
+	compileOnly project(":api-bukkit")
+	compileOnly(libs.paper.api)
+	compileOnly(libs.pluginannotations)
+	annotationProcessor(libs.pluginannotations)
+}
