+++ rewind_v1_5/rewind_v1_5-platform-bungee/build.gradle
@@ -1,0 +1,12 @@
+
+plugins {
+	id "java-library"
+}
+
+dependencies {
+	implementation project(":rewind_v1_5:rewind_v1_5-common")
+	compileOnly project(":api-bungee")
+	compileOnly(libs.bungeecord.api)
+	compileOnly(libs.pluginannotations)
+	annotationProcessor(libs.pluginannotations)
+}
