+++ settings.gradle
@@ -18,6 +18,10 @@ include "core:core-platform-velocity"
 include "core-config"
 include "protocol-game"
 include "rewind_v1_5"
+include "rewind_v1_5:rewind_v1_5-common"
+include "rewind_v1_5:rewind_v1_5-platform-bukkit"
+include "rewind_v1_5:rewind_v1_5-platform-bungee"
+include "rewind_v1_5:rewind_v1_5-platform-velocity"
 include "skin-cache"
 include "stubs"
 include "eaglermotd"
